better handling of grid coordinates

+++ src/game/components.rs
@@ -11,21 +11,21 @@ pub struct Flag {
 #[derive(Component)]
 pub struct Visible;
 
-#[derive(Component, Eq, PartialEq, Debug, Clone)]
+#[derive(Component, PartialEq, Debug, Clone)]
 pub struct Cell {
-    pub(crate) x: u32,
-    pub(crate) y: u32,
+    pub(crate) x: f32,
+    pub(crate) y: f32,
 }
 
 impl Cell {
-    pub fn new(x: u32, y: u32) -> Self {
+    pub fn new(x: f32, y: f32) -> Self {
         Self { x, y }
     }
 
     pub fn is_near(&self, other: &Self) -> bool {
-        (self.x == other.x + 1 || (other.x > 0 && self.x == other.x - 1) || self.x == other.x)
-            && (self.y == other.y + 1
-                || (other.y > 0 && self.y == other.y - 1)
+        (self.x == other.x + 1.0 || (other.x > 0.0 && self.x == other.x - 1.0) || self.x == other.x)
+            && (self.y == other.y + 1.0
+                || (other.y > 0.0 && self.y == other.y - 1.0)
                 || self.y == other.y)
     }
 }
